# Latency optimisation: generating presigned urls asynchronously