--- a/lambda/layers/response_utils_layer/response_utils.py
+++ b/lambda/layers/response_utils_layer/response_utils.py
@@ -1,4 +1,6 @@
 import os
+import asyncio
+import aioboto3
 import boto3
 import logging
 import json
@@ -100,3 +102,36 @@ def generate_presigned_url(bucket: str, key: str, expires_in: int = 3600) -> str
     except Exception as e:
         logger.error(f"Error generating presigned URL: {e}")
         raise
+
+
+async def generate_presigned_url_async(
+    bucket: str,
+    key: str,
+    expires_in: int = 3600,
+    s3_client=aioboto3.Session().client("s3"),
+) -> str:
+    try:
+        url = await s3_client.generate_presigned_url(
+            ClientMethod="get_object",
+            Params={"Bucket": bucket, "Key": key},
+            ExpiresIn=expires_in,
+        )
+
+        return url
+    except Exception as e:
+        logger.error(f"Error generating presigned URL: {e}")
+        raise
+
+
+def add_presigned_urls(items, expires_in):
+    asyncio.run(add_presigned_urls_async(items, expires_in))
+
+
+async def add_presigned_urls_async(items, expires_in):
+    session = aioboto3.Session()
+    async with session.client("s3") as s3_client:  # type: ignore (type error in aioboto3 library)
+        for item in items:
+            if item.get("s3_bucket") and item.get("s3_key"):
+                item["url"] = await generate_presigned_url_async(
+                    item["s3_bucket"], item["s3_key"], expires_in, s3_client
+                )