- Added default value to `expires_in` param for `add_presigned_urls`
- search query http response now includes pagination metadata (page, limit, total)
  - for now, since we are not doing true pagination, total can never be greater than limit
- fetch_videos_handler lambda and search_handler lambdas both use asynchronous processes to generate presigned urls
- ~4.5s -> 2.5s for fetching 25 videos
- ~5.5s -> ~3.5s for fetching 20 search results